requirements.txt
fastapi
uvicorn
python-dotenv
psycopg2-binary
sentence-transformers
requests
pydantic
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-jose
passlib[bcrypt]
python-multipart
